refactor(sql-orm-client): move buildMtiJoins into query-plan-scope
buildMtiJoins moves out of query-plan-select.ts, verbatim, into
query-plan-scope.ts alongside the other row/join-scope helpers query-plan-
select.ts and query-plan-aggregate.ts both already share from there.
query-plan-select.ts imports it back; the SQL it compiles is unchanged.
Reusing it across two exporting modules is the coupling query-plan-scope
exists to remove — the D2 module is not narrowed to row-scope machinery
specifically, so it is the shared home rather than a third module.
