+++ packages/3-extensions/sql-orm-client/src/query-plan-aggregate.ts
@@ -23,8 +23,7 @@ import { plainAggregateExpr, resolveAggregate } from './aggregate-codecs';
 import { resolvePolymorphismInfo } from './collection-contract';
 import { ormError } from './orm-errors';
 import { buildOrmQueryPlan, deriveParamsFromAst } from './query-plan-meta';
-import { buildStateWhere, wrapWithRowNumberDedup } from './query-plan-scope';
-import { buildMtiJoins } from './query-plan-select';
+import { buildMtiJoins, buildStateWhere, wrapWithRowNumberDedup } from './query-plan-scope';
 import { tableSourceForContract } from './storage-resolution';
 import type { AggregateSelector, CollectionState } from './types';
 import { combineWhereExprs } from './where-utils';
